requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.7.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.3.1
babel==2.16.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
Brlapi==0.8.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.0
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.11
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
commonmark==0.9.1
contourpy==1.3.0
coverage==5.5
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.6.7
dbus-python==1.2.18
debugpy==1.8.5
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.7
duplicity==0.8.21
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.14
fastavro==1.10.0
fasteners==0.14.1
fastjsonschema==2.20.0
filelock==3.6.0
Flask==2.1.3
Flask-Alembic==2.0.1
Flask-Cors==5.0.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
flatbuffers==24.12.23
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
future==0.18.2
gast==0.4.0
geckodriver-autoinstaller==0.1.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.74.0
grpcio-status==1.71.2
gTTS==2.5.3
gunicorn==23.0.0
gyp==0.1
h11==0.16.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.20.2
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kazam==1.4.5
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.7
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.46
langchain-text-splitters==0.3.7
langsmith==0.3.18
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
llvmlite==0.44.0
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
MarkupSafe==2.0.1
marshmallow==3.26.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mediapipe==0.10.5
mistune==3.0.2
ml-dtypes==0.4.0
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
namex==0.0.8
narwhals==1.23.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.46
openai==0.28.0
openai-whisper==20240930
opencv-contrib-python==4.6.0.66
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.12.1
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
pillow==11.1.0
platformdirs==2.6.2
pluggy==1.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycairo==1.20.1
pycodestyle==2.12.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.7
pydantic-settings==2.8.1
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==1.0.1
pyscript==0.2.5
pyshorteners==1.0.1
PySocks==1.7.1
pytest==6.2.5
pytest-cov==2.12.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.1
python-json-logger==2.0.7
python3-xlib==0.15
pytweening==1.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rich-click==1.8.2
rpds-py==0.20.0
rsa==4.9
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.1
selenium==4.25.0
Send2Trash==1.8.3
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soupsieve==2.6
split-folders==0.5.1
SQLAlchemy==1.4.23
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
streamlit==1.41.1
sympy==1.13.1
systemd-python==234
tenacity==9.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
triton==3.1.0
typer==0.4.2
types-python-dateutil==2.9.0.20241003
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
usb-creator==0.3.7
uvicorn==0.35.0
virtualenv==20.13.0+ds
voila==0.5.7
wadllib==1.3.6
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.0.1
whitenoise==6.9.0
widgetsnbextension==4.0.13
wrapt==1.17.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.18.3
zipp==1.0.0
zstandard==0.23.0
